// src/@types/pokemon.d.ts
// on a généré les interfaces avec l'aide du site :
// https://transform.tools/json-to-typescript
export interface Name {
  fr: string
  en: string
  jp: string
}
export interface Sprites {
  regular: string
  shiny: string
  gmax: any
}
export interface Type {
  name: string
  image: string
}
export interface Talent {
  name: string
  tc: boolean
}
export interface Stats {
  hp: number
  atk: number
  def: number
  spe_atk: number
  spe_def: number
  vit: number
}
export interface Resistance {
  name: string
  multiplier: number
}
export interface Pre {
  pokedexId: number
  name: string
}
export interface Sprites2 {
  regular: string
  shiny: string
}
export interface Mega {
  orbe: string
  sprites: Sprites2
}
export interface Sexe {
  male: number
  female: number
}
export interface Next {
  pokedexId: number
  name: string
  condition: string
}
export interface Evolution {
  pre: Pre[]
  next: Next[]
  mega: Mega[]
}
export interface Pokemon {
  pokedexId: number
  generation: number
  category: string
  name: Name
  sprites: Sprites
  types: Type[]
  talents: Talent[]
  stats: Stats
  resistances: Resistance[]
  evolution: Evolution
  height: string
  weight: string
  egg_groups: string[]
  sexe: Sexe
  catch_rate: number
  level_100: number
  forme: any
}